+++ CHANGELOG.md
@@ -19,3 +19,6 @@
 - Security checks added.
 - Added SQL storage.
 - Refactored `Request` class name to `AccessRequest`. The name `Request` still supported for backward compatibility. 
+
+# v0.3.1
+- Fixed import dependency error for storage. Udated import statements from `py_abac/storage/__init__.py`. Thanks [dylanmcreynolds](https://github.com/dylanmcreynolds) for PR.